convert project to poetry. Fix some errors
